+++ relay-config/src/config.rs
@@ -653,6 +653,12 @@ pub struct Limits {
     ///
     /// By default there is no idle timeout.
     pub idle_timeout: Option<u64>,
+    /// Sets the maximum number of concurrent connections.
+    ///
+    /// Upon reaching the limit, the server will stop accepting connections.
+    ///
+    /// By default there is no limit.
+    pub max_connections: Option<usize>,
     /// The TCP listen backlog.
     ///
     /// Configures the TCP listen backlog for the listening socket of Relay.
@@ -691,6 +697,7 @@ impl Default for Limits {
             shutdown_timeout: 10,
             keepalive_timeout: 5,
             idle_timeout: None,
+            max_connections: None,
             tcp_listen_backlog: 1024,
         }
     }
@@ -2319,6 +2326,11 @@ impl Config {
     /// Returns the server idle timeout in seconds.
     pub fn idle_timeout(&self) -> Option<Duration> {
         self.values.limits.idle_timeout.map(Duration::from_secs)
+    }
+
+    /// Returns the maximum connections.
+    pub fn max_connections(&self) -> Option<usize> {
+        self.values.limits.max_connections
     }
 
     /// TCP listen backlog to configure on Relay's listening socket.