+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 # Changelog
 
 ## Unreleased
-- Add data categories for LogItem and LogByte. ([#4448](https://github.com/getsentry/relay/pull/4455))
+
+**Features**:
+
+- Add configuration option to limit the amount of concurrent http connections. ([#4453](https://github.com/getsentry/relay/pull/4453))
+
+**Internal**:
+
+- Add data categories for LogItem and LogByte. ([#4455](https://github.com/getsentry/relay/pull/4455))
 
 ## 25.1.0
 
-**Internal**
+**Internal**:
 
 - Updates performance score calculation on spans and events to also store cdf values as measurements. ([#4438](https://github.com/getsentry/relay/pull/4438))
 - Allow log ingestion behind a flag, only for internal use currently. ([#4448](https://github.com/getsentry/relay/pull/4448))